+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 3.56.0
+## 3.57.0 
+* Forward `processor_comments` to `forwarded_comments`
+* Add Venmo 'TokenIssuance' gateway rejection reason
+* Add `AmountNotSupportedByProcessor` to validation error
+
+## 3.56.0
 * Add PayPalHere details
 * Add `networkResponseCode` and `networkResponseText` to transactions and verifications
 * Add `cavv`, `xid`, `ds_transaction_id`, `eci_flag`, and `three_d_secure_version`, to `three_d_secure_info`